Refactor table and restaurant models: remove old model files, update routes and service logic, and enhance error handling
