+++ routes/tableRoutes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { getAllTablesController, getTableByIdController, createTableController, updateTableController, deleteTableController } = require('../controller/tableController');
+const { getAllTablesController, getTableByIdController, createTableController, updateTableController, deleteTableController } = require('../controllers/tableController');
 
 router.get('/', getAllTablesController);
 router.get('/:id', getTableByIdController);
