--- util/AppError.js
@@ -1,9 +1,0 @@
-class AppError extends Error {
-    constructor(statusCode, message) {
-        super(message);
-        this.statusCode = statusCode;
-        Error.captureStackTrace(this, this.constructor); // Capture stack trace for debugging
-    }
-}
-
-module.exports = AppError;
